+++ src/vec2.rs
@@ -17,7 +17,7 @@ use crate::{
 /// any tuple-like object where vector math operations are benefit.
 ///
 /// It is meant to be as convinient as possible to work with in many use cases.
-#[derive(Debug, Copy, Clone, Default, PartialEq, Eq)]
+#[derive(Debug, Copy, Clone, Default, PartialEq, Eq, Hash)]
 pub struct Vec2<T = usize> {
     /// The first coordinate of the vector (x, w, [0]).
     pub x: T,